refactor: unify route auth imports
