+++ server/db.ts
@@ -2,8 +2,8 @@
 import { createClient } from '@supabase/supabase-js';
 import { drizzle } from 'drizzle-orm/postgres-js';
 import postgres from 'postgres';
-import * as schema from "@shared/schema";
-import * as superadminSchema from "@shared/superadmin-schema";
+import * as schema from "../shared/src/schema.js";
+import * as superadminSchema from "../shared/superadmin-schema.js";
 
 if (!process.env.SUPABASE_URL || !process.env.SUPABASE_SERVICE_ROLE_KEY) {
   throw new Error(
