+++ server/routes/admin-analytics.ts
@@ -2,7 +2,7 @@ import type { Express } from "express";
 import { db } from "../db";
 import { users, organizations, trips, activities } from "@shared/schema";
 import { eq, count, sql, and, gte, desc } from "drizzle-orm";
-import { validateJWT } from '../middleware/jwtAuth';
+import { authenticate as validateJWT } from '../middleware/secureAuth.js';
 import { injectOrganizationContext, validateOrganizationAccess } from '../middleware/organizationContext';
 
 interface AdminAnalytics {
