+++ server/routes/ai.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { validateJWT } from "../middleware/jwtAuth";
+import { authenticate as validateJWT } from '../middleware/secureAuth.js';
 import { injectOrganizationContext, validateOrganizationAccess } from '../middleware/organizationContext';
 import { z } from "zod";
 import OpenAI from "openai";
