+++ server/routes/corporateCards.ts
@@ -2,7 +2,7 @@ import type { Express } from "express";
 import { db } from "../db";
 import { corporateCards, cardTransactions, users } from "@shared/schema";
 import { eq, and, desc } from "drizzle-orm";
-import { validateJWT } from '../middleware/jwtAuth';
+import { authenticate as validateJWT } from '../middleware/secureAuth.js';
 import { injectOrganizationContext, validateOrganizationAccess } from '../middleware/organizationContext';
 
 export function registerCorporateCardRoutes(app: Express) {
