+++ server/routes/expenses.ts
@@ -3,7 +3,7 @@ import { eq, and, gte, lte, sql } from 'drizzle-orm';
 import { desc } from 'drizzle-orm';
 import { db } from '../db';
 import { expenses, trips, users } from '@shared/schema';
-import { validateJWT } from '../middleware/jwtAuth';
+import { authenticate as validateJWT } from '../middleware/secureAuth.js';
 import { injectOrganizationContext, validateOrganizationAccess } from '../middleware/organizationContext';
 import { z } from 'zod';
 import { approvalEngine } from '../approvalEngine';
