+++ server/routes/payments.ts
@@ -6,7 +6,8 @@ import { invoices } from '../db/invoiceSchema';
 import { billingEvents } from '../db/superadminSchema';
 import { eq } from 'drizzle-orm';
 import nodemailer from 'nodemailer';
-import { authenticateJWT, type AuthRequest } from '../middleware/auth';
+import { authenticate as authenticateJWT } from '../middleware/secureAuth.js';
+import type { AuthenticatedRequest as AuthRequest } from '../src/types/auth-user.js';
 import { injectOrganizationContext, validateOrganizationAccess } from '../middleware/organizationContext';
 import { logger } from '../utils/logger';
 import type { Invoice } from '../db/db';
