+++ server/routes/subscription-status.ts
@@ -1,5 +1,5 @@
 import { Router, Request, Response } from 'express';
-import { validateJWT } from '../middleware/jwtAuth';
+import { authenticate as validateJWT } from '../middleware/secureAuth.js';
 import { getSubscriptionStatus, checkTripLimit, checkUserLimit } from '../middleware/subscription-limits';
 
 const router = Router();
