+++ server/services/database.service.ts
@@ -1,7 +1,7 @@
 import { Pool, PoolConfig } from 'pg';
 import { createClient } from 'redis';
 import { drizzle } from 'drizzle-orm/node-postgres';
-import * as schema from "@shared/schema";
+import * as schema from "../../shared/src/schema.js";
 import config from '../config.js';
 import { logger } from '../utils/logger.js';
